HW1/hw1/src/main/java/tqs/homework/hw1/services/CityService.java
package tqs.homework.hw1.services;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import tqs.homework.hw1.models.City;
import tqs.homework.hw1.repositories.CityRepository;

@Service
public class CityService {

    private final CityRepository cityRepository;

    @Autowired
    public CityService(CityRepository cityRepository) {
        this.cityRepository = cityRepository;
    }

    public City getCityById(Long id) {
        return cityRepository.findById(id).orElse(null);
    }

    public City saveCity(City city) {
        return cityRepository.save(city);
    }

    public void deleteCity(City city) {
        cityRepository.delete(city);
    }

    public void deleteCityById(Long id) {
        cityRepository.deleteById(id);
    }

    public List<City> getAllCities() {
        return cityRepository.findAll();
    }

    public City getCityByName(String name) {
        return cityRepository.findByName(name);
    }
    
}
